+++ libs/pavex/src/web/mod.rs
@@ -2,13 +2,12 @@ pub use app::App;
 pub use diagnostic::CompilerDiagnostic;
 
 mod app;
-mod application_state_call_graph;
+mod call_graph;
 mod codegen;
 mod codegen_utils;
 mod constructors;
 pub(crate) mod dependency_graph;
 mod diagnostic;
 mod generated_app;
-mod handler_call_graph;
 mod resolvers;
 mod traits;
